requirements.txt
SpeechRecognition
google-api-python-client
isodate
google-generativeai
SpeechRecognition[audio]
